ADD getter and setter for deviceType

+++ Models/GPU.php
@@ -1,6 +1,6 @@
 <?php
   class GPU {
-    public function __construct(public string $brand, public string $model, public int $frequency, public int $reservedMemory) {
+    public function __construct(public string $brand, public string $model, public float $frequency, public int $reservedMemory) {
       $this -> brand = $brand;
       $this -> model = $model;
       $this -> frequency = $frequency;
